+++ ethergo/listener/listener.go
@@ -55,6 +55,7 @@ type chainListener struct {
 	blockWait       uint64
 	// otelRecorder is the recorder for the otel metrics.
 	otelRecorder iOtelRecorder
+	name         string
 }
 
 var (
@@ -96,7 +97,7 @@ func (c *chainListener) Listen(ctx context.Context, handler HandleLog) (err erro
 	}
 
 	if c.otelRecorder == nil {
-		c.otelRecorder, err = newOtelRecorder(c.handler, int(c.chainID))
+		c.otelRecorder, err = newOtelRecorder(c.handler, int(c.chainID), c.name)
 		if err != nil {
 			return fmt.Errorf("could not create otel recorder: %w", err)
 		}